requirements.txt
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.1
idna==3.7
numpy==2.0.1
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.43.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
